+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 ## [4.0 Snapshot 4] - UNRELEASED
+### New Features
+#### Blocks
+- Added Unlit Soul Campfire
+  - Is given when the player picks up an unlit Soul Campfire block with Silk Touch
+
 ### Changes
+#### Blocks
+- Unlit campfires are now given when the player picks up an unlit Campfire block with Silk Touch
+
 #### Loot
 - Changed cows, mooshrooms, horses, llamas, mules, donkeys, and trader llamas to drop between 1 and 4 leather